+++ modules/core-api/src/main/java/com/enonic/wem/api/xml/mapper/XmlRegionDescriptorMapper.java
@@ -2,18 +2,27 @@ package com.enonic.wem.api.xml.mapper;
 
 import java.util.List;
 
-import com.google.common.collect.Lists;
-
 import com.enonic.wem.api.content.page.region.RegionDescriptor;
 import com.enonic.wem.api.content.page.region.RegionDescriptors;
 import com.enonic.wem.api.xml.model.XmlRegionDescriptor;
+import com.enonic.wem.api.xml.model.XmlRegions;
 
 final class XmlRegionDescriptorMapper
 {
-    protected static RegionDescriptors fromXml( final List<XmlRegionDescriptor> regions )
+    protected static RegionDescriptors fromXml( final XmlRegions regions )
     {
         final RegionDescriptors.Builder builder = RegionDescriptors.newRegionDescriptors();
 
+        if ( regions != null )
+        {
+            fromXml( builder, regions.getRegion() );
+        }
+
+        return builder.build();
+    }
+
+    private static void fromXml( final RegionDescriptors.Builder builder, final List<XmlRegionDescriptor> regions )
+    {
         if ( regions != null )
         {
             for ( final XmlRegionDescriptor descriptor : regions )
@@ -21,8 +30,6 @@ final class XmlRegionDescriptorMapper
                 builder.add( fromXml( descriptor ) );
             }
         }
-
-        return builder.build();
     }
 
     protected static RegionDescriptor fromXml( final XmlRegionDescriptor descriptor )
@@ -32,12 +39,12 @@ final class XmlRegionDescriptorMapper
         return builder.build();
     }
 
-    protected static List<XmlRegionDescriptor> toXml( final RegionDescriptors object )
+    protected static XmlRegions toXml( final RegionDescriptors object )
     {
-        final List<XmlRegionDescriptor> result = Lists.newArrayList();
+        final XmlRegions result = new XmlRegions();
         for ( final RegionDescriptor descriptor : object )
         {
-            result.add( toXml( descriptor ) );
+            result.getRegion().add( toXml( descriptor ) );
         }
         return result;
     }