+++ modules/core-api/src/main/java/com/enonic/wem/api/xml/serializer/XmlPageDescriptorSerializer.java
@@ -1,0 +1,21 @@
+package com.enonic.wem.api.xml.serializer;
+
+import javax.xml.bind.JAXBElement;
+
+import com.enonic.wem.api.xml.model.XmlPageComponentElem;
+import com.enonic.wem.api.xml.model.XmlPageDescriptor;
+
+final class XmlPageDescriptorSerializer
+    extends XmlSerializerBase<XmlPageDescriptor>
+{
+    public XmlPageDescriptorSerializer()
+    {
+        super( XmlPageDescriptor.class );
+    }
+
+    @Override
+    protected JAXBElement<XmlPageDescriptor> wrap( final XmlPageDescriptor value )
+    {
+        return new XmlPageComponentElem( value );
+    }
+}